+++ CHANGELOG.md
@@ -5,11 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Removed all traits from collision API. Duplicate methods now have suffixes
+  that match the parameter types. For example,
+  `BoundingBox::Contains(&BoundingSphere)` and `BoundingBox::Contains(&BoundingBox)` are now
+  `BoundingBox::ContainsSphere(&BoundingSphere)` and `BoundingBox::ContainsBox(&BoundingBox)`.
 
 ## [0.1.0-alpha.2] - 2020-05-07
 ### Added
 - Documentation updates
-- Added collision API
+- Collision API
 
 ### Changed
 - XMMatrixInverse now accepts `None` for the determinant, to match the upstream API